Game network messages have been moved to core headers #403
